pyproject.toml
[project]
name = "scryfall-download"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "contxt>=0.1.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.8.0",
    "pytest>=7.4.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[tool.uv.workspace]
members = ["src/mcp"]
